# services/design-frames-service/openapi.yaml
openapi: 3.0.3
info:
  title: design-frames-service
  version: 0.2.0
  description: >
    Product-design-phase service: navigable HTML frames, per-flow approval,
    and the API/component/flag contract seam. Extracted from FuzeFront's
    design/frames/** pipeline — see README.md and docs/EXTRACTION.md.


    v0.2.0 introduces the Postgres-backed lifecycle tier (contract only; the
    Express/pg runtime lands in a later stream). Frame CONTENT stays in git
    (the owning product's design/frames/<feature>/); Postgres holds ONLY
    lifecycle/index metadata + refs, keyed by content-stamp. Entities use the
    product-local `fxdf` identifier namespace (see .fuze/manifest.json and
    docs/postgres-tier.md). The entire v0.1.0 /api/v1/features/** surface is
    unchanged and backward-compatible — the new tier projects it from the new
    model. See docs/postgres-tier.md for the two-tier model and migration path.
  x-changelog:
    - version: 0.2.0
      date: '2026-08-19'
      changes:
        - 'ADD /api/v1/projects CRUD (create mints fxdf_prj_*, additionalProperties:false, no client id) + GET /{id}/features.'
        - 'ADD optional projectId reference on POST /api/v1/features (reference, not identity).'
        - 'ADD GET /api/v1/features/{slug}/flows/{flowId}/approvals — append-only decision history (paginated).'
...
    - version: 0.3.0
      date: '2026-10-09'
      changes:
...
    - version: 0.1.0
      date: '2026-08-11'
      changes:
...
servers:
  - url: http://localhost:4400
security:
  - bearerAuth: []
    delegatedUser: []
  - bearerAuth: []
  - {}
paths:
...
  /api/v1/features:
    get:
      summary: List features
      security: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  features:
                    type: array
                    items: { $ref: '#/components/schemas/FeatureSummary' }
    post:
      summary: Create a feature shell
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  slug: { type: string }
                  manifest: { $ref: '#/components/schemas/Manifest' }
        '400': { $ref: '#/components/responses/Validation' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [slug]
              properties:
                slug: { type: string, pattern: '^[a-z0-9]+(-[a-z0-9]+)*$' }
                name: { type: string }
                description: { type: string }
                designSystem: { type: string }
                entry: { type: string }
                sourceRepo: { type: string, nullable: true, description: "Which consuming product these frames are for, e.g. 'FuzeFront'." }
                projectId:
                  allOf: [ { $ref: '#/components/schemas/ProjectId' } ]
                  nullable: true
                  description: >
                    OPTIONAL reference to an existing project (fxdf_prj_*) this
                    feature belongs to. A reference to an entity that already
                    exists, NOT identity (identifier-standard §1): the feature's
                    own id is still minted server-side. Omitted/null = unassigned.
  /api/v1/features/{slug}:
    get:
      summary: Get a feature's manifest and all frame contents
      security: []
      parameters: [{ $ref: '#/components/parameters/Slug' }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  slug: { type: string }
                  manifest: { $ref: '#/components/schemas/Manifest' }
                  frames:
                    type: object
                    additionalProperties: { type: string }
                    description: frame file name -> raw HTML content
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/features/{slug}/manifest:
    put:
      summary: Replace the manifest (schema-validated)
      parameters: [{ $ref: '#/components/parameters/Slug' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Manifest' }
      responses:
        '200': { description: OK }
        '400': { $ref: '#/components/responses/Validation' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/features/{slug}/import:
    post:
      summary: Atomically publish a complete repository frame set and immutable revision
      description: >
        Replaces the entire current manifest and frame set in one publication,
        removing files absent from the request. expectedStamp is optional for an
        unconditional publish, null for create-only, or the observed current
        content stamp for compare-and-swap. A mismatch returns 409. Successful
        retries reuse the same immutable revision and repair missing frame indexes.
      parameters: [{ $ref: '#/components/parameters/Slug' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [manifest, frames]
              properties:
                manifest: { $ref: '#/components/schemas/Manifest' }
                frames:
                  type: object
                  description: Complete flat HTML filename-to-content map; entry and all declared frames are required.
                  additionalProperties: { type: string }
                expectedStamp:
                  type: string
                  nullable: true
                  pattern: '^[0-9a-f]{64}$'
      responses:
        '200':
          description: Complete imported revision published
          content:
            application/json:
              schema:
                type: object
                required: [slug, stamp, frameCount, revision]
                properties:
                  slug: { type: string }
                  stamp: { type: string, pattern: '^[0-9a-f]{64}$' }
                  frameCount: { type: integer, minimum: 0 }
                  revision: { $ref: '#/components/schemas/ContentRevision' }
        '400': { $ref: '#/components/responses/Validation' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/StampConflict' }
  /api/v1/features/{slug}/stamp:
    get:
      summary: Compute the current content stamp and compare to the persisted one (no write)
      security: []
      parameters: [{ $ref: '#/components/parameters/Slug' }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  slug: { type: string }
                  stamp: { type: string, pattern: '^[0-9a-f]{64}$' }
                  manifestStamp: { type: string, nullable: true }
                  current: { type: boolean }
    post:
      summary: Atomically compute, snapshot, and persist the exact current content stamp
      parameters: [{ $ref: '#/components/parameters/Slug' }]
      responses:
        '200': { description: OK }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/features/{slug}/revisions:
    get:
      summary: List immutable content revisions for a feature
      description: >
        Each item represents the exact frame bytes committed under a unique
        content stamp. Revisions are created by full imports, review decisions,
        or POSTing the feature stamp and
        are never updated in place.
      security: []
      parameters: [{ $ref: '#/components/parameters/Slug' }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [slug, revisions]
                properties:
                  slug: { type: string }
                  revisions:
                    type: array
                    items: { $ref: '#/components/schemas/ContentRevision' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/features/{slug}/revisions/{stamp}:
    get:
      summary: Retrieve the exact manifest and frame bytes for an immutable revision
      security: []
      parameters:
        - { $ref: '#/components/parameters/Slug' }
        - name: stamp
          in: path
          required: true
          schema: { type: string, pattern: '^[0-9a-f]{64}$' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [slug, stamp, revision, manifest, frames]
                properties:
                  slug: { type: string }
                  stamp: { type: string, pattern: '^[0-9a-f]{64}$' }
                  revision: { $ref: '#/components/schemas/ContentRevision' }
                  manifest: { $ref: '#/components/schemas/Manifest' }
                  frames:
                    type: object
                    additionalProperties: { type: string }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/features/{slug}/revisions/{stamp}/frame-refs:
    get:
      summary: Resolve frame identities within one immutable revision
      description: >
        Returns only frame references indexed for the supplied content stamp.
        Use the returned frame id as targetRef for frame or element discussions;
        this prevents annotations from being attached to a newer frame with the
        same file name.
      security: []
      parameters:
        - { $ref: '#/components/parameters/Slug' }
        - name: stamp
          in: path
          required: true
          schema: { type: string, pattern: '^[0-9a-f]{64}$' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [slug, stamp, frames]
                properties:
                  slug: { type: string }
                  stamp: { type: string, pattern: '^[0-9a-f]{64}$' }
                  frames:
                    type: array
                    items: { $ref: '#/components/schemas/RevisionFrameRef' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/features/{slug}/frames/{file}:
    parameters:
      - { $ref: '#/components/parameters/Slug' }
      - name: file
        in: path
        required: true
        schema: { type: string, pattern: '\.html$' }
    get:
      summary: Get one frame's raw HTML
      security: []
      responses:
        '200': { description: OK }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      summary: Create/replace one frame's HTML content
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [html]
              properties:
                html: { type: string }
      responses:
        '200': { description: OK }
        '400': { $ref: '#/components/responses/Validation' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      summary: Delete a frame
      responses:
        '204': { description: No Content }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/features/{slug}/flows/{flowId}/approve:
    post:
      summary: Approve a flow (appends an approve decision to the flow's log)
      description: >
        Appends an `approve` decision to the flow's append-only approval log
        (never mutates an existing row). Backward-compatible: an old caller
        sending only `{ approvedBy }` still works and binds the decision to the
        current content stamp under the write guard. If `contentStamp` is
        supplied it binds the decision to the exact frames the reviewer saw
        (must equal the feature's current stamp, else 409). The response echoes
        the created approval (fxdf_apr_*) and the stamp it was bound to.
      parameters:
        - { $ref: '#/components/parameters/Slug' }
        - name: flowId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [approvedBy]
              properties:
                approvedBy: { type: string, description: 'Legacy compatibility field. Persisted actor is always the verified delegated user or machine subject.' }
                actorType:
                  type: string
                  nullable: true
                  enum: [user, agent]
                  description: 'Legacy compatibility field. Server derives user for delegated users and agent for machines.'
                contentStamp:
                  type: string
                  pattern: '^[0-9a-f]{64}$'
                  description: 'Sha256 stamp the reviewer saw. Omission binds to current content; null or malformed values fail validation. Mismatch with current content returns 409.'
      responses:
        '200':
          description: Approve decision recorded
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Approval' }
        '400': { $ref: '#/components/responses/Validation' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/StampConflict' }
  /api/v1/features/{slug}/flows/{flowId}/reject:
    post:
      summary: Reject a flow (appends a reject decision to the flow's log)
      description: >
        Appends a `reject` decision (with a required `reason`) to the flow's
        append-only approval log. Reject re-dispatches product-designer for an
        improving iteration; it does not delete prior rows. The response echoes
        the created approval (fxdf_apr_*) and its bound stamp. Omission of
        contentStamp binds to current content under the write guard.
      parameters:
        - { $ref: '#/components/parameters/Slug' }
        - name: flowId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [reason]
              properties:
                reason: { type: string, minLength: 1, description: 'Why the flow was rejected. REQUIRED as of v0.2.0 — an append-only decision log row must carry its rationale.' }
                rejectedBy: { type: string, nullable: true, description: 'Legacy compatibility field. Persisted actor is always the verified delegated user or machine subject.' }
                actorType:
                  type: string
                  nullable: true
                  enum: [user, agent]
                  description: 'Legacy compatibility field. Server derives user for delegated users and agent for machines.'
                contentStamp:
                  type: string
                  pattern: '^[0-9a-f]{64}$'
                  description: 'Sha256 stamp the reviewer saw; omission binds to current content. Null and malformed values fail validation; stale stamps return 409.'
      responses:
        '200':
          description: Reject decision recorded
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Approval' }
        '400': { $ref: '#/components/responses/Validation' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/features/{slug}/flows/{flowId}/approvals:
    get:
      summary: List a flow's append-only approval decision history (newest first)
      security: []
      parameters:
        - { $ref: '#/components/parameters/Slug' }
        - name: flowId
          in: path
          required: true
          schema: { type: string }
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Cursor' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [items, page]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Approval' }
                  page: { $ref: '#/components/schemas/PageInfo' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/v1/projects/{id}/features:
    get:
      summary: List the features belonging to a project (paginated)
      security: []
      parameters:
        - { $ref: '#/components/parameters/ProjectIdPath' }
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Cursor' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [items, page]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/FeatureSummary' }
                  page: { $ref: '#/components/schemas/PageInfo' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    delegatedUser:
      type: apiKey
      in: header
      name: X-Fuze-Delegation
      description: >
        Server-to-server header with value Bearer <delegation-token>. FuzeFront's
        same-origin proxy exchanges the browser session for a fuze-delegation
        token with audience service:fuzex. Authorization must contain its
        immediate actor's fuze-workload token. Both are introspected; actor.sub
        must match the workload subject and tenant-bound workloads must match
        the delegated tenant. Delegation requires fuzex:frames:read for reads
        and fuzex:frames:write for mutations. Invalid supplied delegation never
        falls back to anonymous reads or direct machine writes. Browsers never
        receive either upstream credential.
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        A FuzeFront-issued MACHINE token for direct service writes, or the
        fuze-workload token accompanying delegatedUser for hosted user calls.
        Reads without delegation are public. All write methods require auth.

        Obtain one from FuzeFront's `POST /api/v1/security/tokens`. This service
        verifies it per request against FuzeFront's
        `POST /api/v1/security/tokens/introspect` and requires the scope
        `fuzex:frames:write`; a token without it gets 403, not 401.

        Verification is fail-closed: introspection answers HTTP 200 even for an
        inactive token, so the decision is made on the `active` boolean in the
        response body. An unreachable or unusable introspection endpoint denies
        the request rather than allowing it.

        This replaced the `DESIGN_FRAMES_API_TOKENS` pre-shared bearer list
        (issue #26), which is deleted — there is no fallback.
...